Splits charts into their own section
