+++ lib/components/Charts/exports.ts
@@ -1,0 +1,1 @@
+export * from "./Area"
