+++ lib/components/Charts/utils/colors.tsx
