+++ lib/components/Insights/Types/AreaInsight/index.stories.tsx
@@ -1,53 +1,20 @@
 import type { Meta, StoryObj } from "@storybook/react"
 
-import { AreaInsight, AreaInsightProps } from "."
+import AreaChartStory from "@/components/Charts/Area/index.stories"
+import { AreaInsight } from "."
 import ContainerStory from "../../Container/index.stories"
 
-const Component = AreaInsight
-
-const dataConfig = {
-  desktop: {
-    label: "Desktop",
-  },
-  mobile: {
-    label: "Mobile",
-  },
-}
-
 const meta = {
-  component: Component,
+  component: AreaInsight,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
   args: {
     ...ContainerStory.args,
-    chart: {
-      dataConfig,
-      xAxis: {
-        hide: false,
-        tickFormatter: (value: string) => value.slice(0, 3),
-      },
-      yAxis: {
-        hide: true,
-      },
-      data: [
-        { label: "January", values: { mobile: 4000, desktop: 2400 } },
-        { label: "February", values: { mobile: 3000, desktop: 1398 } },
-        { label: "March", values: { mobile: 2000, desktop: 4000 } },
-        { label: "April", values: { mobile: 1500, desktop: 8000 } },
-        { label: "May", values: { mobile: 2000, desktop: 6000 } },
-        { label: "June", values: { mobile: 3500, desktop: 4000 } },
-        { label: "July", values: { mobile: 4500, desktop: 2000 } },
-        { label: "August", values: { mobile: 5500, desktop: 1000 } },
-        { label: "September", values: { mobile: 6500, desktop: 500 } },
-        { label: "October", values: { mobile: 7500, desktop: 300 } },
-        { label: "November", values: { mobile: 8500, desktop: 200 } },
-        { label: "December", values: { mobile: 9500, desktop: 500 } },
-      ],
-    },
-  } satisfies AreaInsightProps<typeof dataConfig>,
-} satisfies Meta<typeof Component>
+    chart: AreaChartStory.args,
+  },
+} satisfies Meta<typeof AreaInsight>
 
 export default meta
 type Story = StoryObj<typeof meta>
