+++ lib/components/Insights/Types/AreaInsight/index.tsx
@@ -1,40 +1,17 @@
 import {
-  ChartConfig,
-  ChartContainer,
-  ChartTooltip,
-  ChartTooltipContent,
-} from "@/ui/chart"
+  AreaChart,
+  AreaChartConfig,
+  AreaChartProps,
+  InferAreaKeys,
+} from "@/components/Charts/Area"
 import { ForwardedRef, forwardRef } from "react"
-import { Area, AreaChart, CartesianGrid, XAxis, YAxis } from "recharts"
 import { InsightsContainer, InsightsContainerProps } from "../../Container"
-import { autoColor } from "../utils/colors"
-
-type Key = string
-
-type ChartItem<AreaKeys extends Key> = {
-  label: string
-  values: Record<AreaKeys, number>
-}
-
-type AxisConfig = {
-  hide?: boolean
-  tickFormatter?: (value: string) => string
-}
-
-type AreaConfig<Keys extends Key = string> = Record<Keys, ChartConfig[string]>
-
-type InferAreaKeys<T> = T extends AreaConfig<infer K> ? K : never
 
 export type AreaInsightProps<
-  DataConfig extends AreaConfig = AreaConfig,
-  AreaKeys extends Key = InferAreaKeys<DataConfig>,
+  DataConfig extends AreaChartConfig = AreaChartConfig,
+  AreaKeys extends string = InferAreaKeys<DataConfig>,
 > = InsightsContainerProps & {
-  chart: {
-    dataConfig: AreaConfig<AreaKeys>
-    data: ChartItem<AreaKeys>[]
-    xAxis?: AxisConfig
-    yAxis?: AxisConfig
-  }
+  chart: AreaChartProps<DataConfig, AreaKeys>
 }
 
 function fixedForwardRef<T, P>(
@@ -46,59 +23,15 @@ function fixedForwardRef<T, P>(
 }
 
 export const _AreaInsight = <
-  DataConfig extends AreaConfig,
-  Keys extends Key = string,
+  DataConfig extends AreaChartConfig,
+  Keys extends string = string,
 >(
-  {
-    chart: { data, dataConfig, xAxis, yAxis },
-    ...containerProps
-  }: AreaInsightProps<DataConfig, Keys>,
+  { chart, ...containerProps }: AreaInsightProps<DataConfig, Keys>,
   ref: ForwardedRef<HTMLDivElement>
 ) => {
-  const areas = Object.keys(dataConfig) as Array<keyof typeof dataConfig>
-
   return (
-    <InsightsContainer {...containerProps}>
-      <ChartContainer config={dataConfig} ref={ref}>
-        <AreaChart
-          accessibilityLayer
-          data={data.map((item) => ({ x: item.label, ...item.values }))}
-          margin={{ left: 12, right: 12 }}
-        >
-          <CartesianGrid vertical={false} />
-          {!xAxis?.hide && (
-            <XAxis
-              dataKey="x"
-              tickLine={false}
-              axisLine={false}
-              tickMargin={8}
-              tickFormatter={xAxis?.tickFormatter}
-            />
-          )}
-          {!yAxis?.hide && (
-            <YAxis
-              tickLine={false}
-              axisLine={false}
-              tickMargin={8}
-              tickFormatter={yAxis?.tickFormatter}
-            />
-          )}
-          <ChartTooltip
-            cursor
-            content={<ChartTooltipContent indicator="dot" />}
-          />
-          {areas.map((area, index) => (
-            <Area
-              key={area}
-              dataKey={area}
-              type="natural"
-              fill={dataConfig[area].color || autoColor(index)}
-              fillOpacity={0.4}
-              stroke={dataConfig[area].color || autoColor(index)}
-            />
-          ))}
-        </AreaChart>
-      </ChartContainer>
+    <InsightsContainer {...containerProps} ref={ref}>
+      <AreaChart {...chart} />
     </InsightsContainer>
   )
 }
