+++ lib/components/exports.ts
@@ -1,5 +1,6 @@
 export * from "./Actions/exports"
 export * from "./Blocks/exports"
+export * from "./Charts/exports"
 export * from "./Forms/exports"
 export * from "./Information/exports"
 export * from "./Insights/exports"
